aware-agent-backend/requirements.txt
# Core dependencies
langchain==0.0.352
langchain-openai>=0.0.2
langchain-core>=0.1.4,<0.2.0
pydantic>=1.7.4,!=1.8,!=1.8.1,<2.0.0
python-dotenv>=0.21.0
fastapi>=0.95.2,<0.100.0
uvicorn>=0.18.3

# NLP and semantic analysis
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
scikit-learn>=1.3.0
numpy>=1.24.0

# Async and concurrency
aiohttp==3.9.1
asyncio==3.4.3

# Database and storage
aiosqlite==0.19.0
sqlalchemy>=2.0.0
alembic==1.7.1
asyncpg==0.29.0
aioredis==2.0.1

# Testing
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==2.12.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.11.0
psutil==5.9.0

# Development tools
black==21.9b0
flake8==3.9.2
mypy==0.910
pylint==3.0.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Additional dependencies
anytree>=2.8.0
beautifulsoup4>=4.9.0
chromadb>=0.4.0
httpx>=0.24.0
markdown>=3.4.0
networkx>=3.0.0
matplotlib>=3.0.0
plotly>=5.0.0
seaborn>=0.12.0
scholarly>=1.7.11
sentence-transformers>=2.5.0
statsmodels>=0.14.0
textstat>=0.7.0
transformers>=4.30.0
wordcloud>=1.9.0
python-louvain>=0.16
leidenalg>=0.9.0
prophet>=1.1.0
tensorflow>=2.15.0
tf-keras>=2.15.0
feedparser>=6.0.0
langchain-community>=0.0.20
websockets==10.1
redis==4.1.0
boto3==1.20.32
pandas>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 